pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "issue-sentinel"
version = "0.1.0"
description = "AI-powered GitHub issue triage — classify, prioritize, and route issues automatically."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Akshay Dixit", email = "kustonaut@github.com"},
]
keywords = ["github", "issues", "triage", "ai", "classification", "nlp", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Bug Tracking",
]

dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
llm = [
    "openai>=1.0",
    "anthropic>=0.18",
]
all = [
    "issue-sentinel[llm]",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.8",
]

[project.scripts]
issue-sentinel = "issue_sentinel.cli:main"

[project.urls]
Homepage = "https://github.com/kustonaut/issue-sentinel"
Repository = "https://github.com/kustonaut/issue-sentinel"
Issues = "https://github.com/kustonaut/issue-sentinel/issues"
Documentation = "https://kustonaut.github.io/issue-sentinel"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
